[ovsp4rt] Extract ovsp4rt_str_to_tunnel_type() (#735)
- Moved ovsp4rt_str_to_tunnel_type() to a separate source
  file for C api functions that are not wrappers.
