+++ ovs-p4rt/sidecar/CMakeLists.txt
@@ -35,10 +35,12 @@ if(BUILD_CLIENT)
     target_sources(ovs_sidecar_o PRIVATE
       ovsp4rt_internal_api.h
       ovsp4rt_journal_api.cc
+      ovsp4rt_simple_api.cc
     )
   else()
     target_sources(ovs_sidecar_o PRIVATE
       ovsp4rt_internal_api.h
+      ovsp4rt_simple_api.cc
       ovsp4rt_standard_api.cc
     )
   endif()
